add comments to lp price function

+++ contracts/peer-to-peer/interfaces/oracles/IUniV2.sol
@@ -4,11 +4,11 @@ pragma solidity 0.8.19;
 interface IUniV2 {
     /**
      * @notice returns reserves of uni v2 pool
-     * @return token0 reserves
-     * @return token1 reserves
+     * @return token0 reserves (use unit256 to avoid overflow if uint112)
+     * @return token1 reserves (use unit256 to avoid overflow if uint112)
      * @return timestamp (which isn't too relevant for us)
      */
-    function getReserves() external view returns (uint112, uint112, uint32);
+    function getReserves() external view returns (uint256, uint256, uint32);
 
     /**
      * @notice token0 address of pool
